v1.1.1 - Add a logger option to print timestamps in UTC for log traces
Write-Log is used for writing all messages to the console and the log file.
It embeds a timestamp with each message, but this timestamp uses LocalTime.
A request has come in to optionally allow users to have this timestamp
recorded in UTC instead of LocalTime.

This adds a new global variable $global:SBUseUTC that defaults to $false.
If set to $true, then Write-Log will use UTC for the timestamps, appending
a "Z" to the end of the timestamp based on http://www.w3.org/TR/NOTE-datetime.

The documentation has been updated to reflect this new
configuration option.

* Additionally fixed a typo in Write-Log where it was trying to
call Write-Debug.

Resolves #10: Add a logger option to print timestamps in UTC for log traces