Fix proxy request encoding and return special headers
The proxy was submitting all requests with the encoding type "utf8".
The correct value for that encoding type is "UTF-8".  The Windows
Store Submission API was robust with the improperly specified encoding
type for some commands, but not others.  This gets the request to be
standard across the board.

Additionally, this updates the proxy to return back all of the special
response headers that the Windows Store Submission API returns.
Most importantly, this is now returning "MS-CorrelationID" which is
used by the Windows Store Submission API team when they are investigating
bug reports.

At this time, StoreBroker itself will never see/report this value because
the PowerShell Invoke-RestMethod (which Invoke-SBRestMethod uses under the
hood) discards all HTTP response headers without any option to prevent
that behavior.  Having these additional headers though will still be useful
if one-off investigations need to be performed.