Clarify behavior of Dev Portal with new submissions
We have a warning on Commit that changes to a submission can't be seen
in the Dev Portal until the submission enters the Certification phase,
however we weren't providing that warning right after a submission was
patched, which was leading to some user confusion of why they couldn't
see their changes.

The warning has been moved to a single location and is now referenced
throughout the module.

Additionally, normalized the capitalization of the phrase "Dev Portal"
throughout the module.