Mandatory Update and Package Rollout Support
Allows users to specify that an application or flight submission
is a mandatory update, or to enable gradual rollout on the packages
for those submissions.

This also introduces new top-level commands to allow users
to further control package rollout (Get / Update / Complete / Stop).

Additionally, the `Format-ApplicationSubmission` and
`Format-ApplicationFlightSubmission` commands were updated to also print
out mandatory update and package rollout information.  No new Format
command was introduced for `Get-ApplicationSubmissionPackageRollout` or
`Get-ApplicationFlightSubmissionPackageRollout` since the returned content
was simple enough as-is.

Additionally:

  * Updated Get-PiiSafeString / Get-SHA512Hash to support receiving
    empty / null strings, as it's possible to call
    `Update-ApplicationSubmission` and `Update-ApplicationFlightSubmission`
    without providing a packagePath, and without this update that would cause
    a parameter binding failure in those scenarios

  * Updated all of the `Patch-*Submission` methods to first convert
    TargetDate to UTC before using it, as that is the expected behavior from
    the API.  The documentation and Comment-Based-Help has been updated to
    indicate to users that the time will be converted to UTC for them.

Resolves #2: Add support for Package Rollout and Mandatory Update