Ensure MS-* headers added to exception responses and telemetry
The previous update that added MS-* headers from the Submission API's
response to the proxy's response did not do the same for the errors.
I've refactored the handling of the HttpWebResponse so that the handling
is done the same for both success and error responses.

Additionally, the MS-CorrelationId header is now added to the telemetry
reported from the Proxy as well to assist with debugging failures.