Prevent "Cannot index into a null array." exception when no headers exist in WebException Response
A PowerShell timeout of a web request will still result in a WebException being thrown,
although there will be no headers in the Response.  If you try to index into a `$null`
value, you end up getting an InvalidOperationException indicating: "Cannot index into a null array."

So, we'll now make sure that `$ex.Response.Headers` has headers before trying to get the
correlationId.  It's ok if the header itself is missing...we'll just get back a $null which
is fine...the issue is only when you try to index into null (e.g. `$null["foo"]`).