Update Invoke-WebRequest to use -UseBasicParsing
StoreBroker is being used on some systems where the
Internet Explorer engine is not available, or Internet
Explorer's first-launch configuration is not complete.

In those scenarios, it's necessary to specify -UseBasicParsing
with Invoke-WebRequest to indicate that it should use the response
object for HTML content without Document Object Model (DOM) parsing.