Fix default NuGet download path for interactive console users
We need to explicitly specify the NuGet source feed for nuget packages,
otherwise we risk donload failure on user machines where they've changed
the default feed configuration.

This fix was done a while ago for users of `-NoStatus`, but it was
accidentally not done for default users.  This corrects that oversight.