Support packaging .appxupload files containing a single .appxbundle.
StoreBroker used to support this scenario but dropped support after documentation indicated an .appxupload should only ever contain a single .appx file, see https://docs.microsoft.com/en-us/windows/uwp/packaging/create-app-package-with-makeappx-tool.

Regardless of the documentation, there are StoreBroker users who have been uploading with this type of file and the Store seems to support it.  This change adds back support for this type of file.