Add -Force switch to Join-SubmissionPackage
Currently, `Join-SubmissionPackage` prevents you from outputting
to a file that already exists. Now with the `-Force` switch, you
can overwrite the same file if you want to.

This was a user request for when they are chaining together multiple
`Join-SubmissionPackage` commands in a row, where the request is to
be able to use the same intermedia file as both the output for one
command, as well as the input (and then output again) of the next
command.