Increment version for NuGet packaging
While there were no code changes to the core module for StoreBroker
with PR #51, there was a change to the XSD that is included in the
NuGet package.  Since the NuGet package version has to increase with
an update, and since we've tied it to the version number of the
module, this version number needs to be incremented.