Clarify how to get AppId ; allow empty strings to be logged
@martinsuchan reported that SETUP.md was a bit confusing
regarding how to get the `AppId`.  I've updated the section
to indicate how to get it from the DevPortal directly, as well
as how to get it with a simplified piping command (without the
`Where-Object`).

Additionally, this removes the restriction on `Write-Log` that
message must be non-null/empty.  Requiring a non-null/empty
string resulted in error messages that focused on bad input
to `Write-Log`, which should really be a black-box command
to users of StoreBroker.  Instead, we'll let empty strings
be logged, which should cause users to look into the command
that they were executing instead.  This is more in-line with
how PowerShell default commands handle situations where there
is no result.

Resolves #63 - Confusing "Getting Your AppId" section in SETUP