Support relative paths that don't begin with '.' to denote the current working directory.
For example, an -OutPath of "SomeFolder" should represent a sub-folder of the current working directory. While PowerShell can handle path fragments of this type, interacting with .NET methods, such as XmlDocument.Save() can cause the file to be saved to an unintended directory.

The fix is to use Resolve-UnverifiedPath in-order to resolve path fragments like this to a full path.

Additonally, adds a new function Ensure-Directory, to ensure a directory is created/exists.

Bump module version to 1.11.0