Enable Language Fallback support for media (screenshots and icons)
The Store Submission API was recently updated to allow for submissions to reference the same media file across multiple elements.  In other words, the same .png file can now be referenced as a screenshot for multiple languages.

Some applications choose not to localize their screenshots, only focusing on localizing the written metadata for their PDP's.  In these scenarios, they would have a smaller StoreBroker package (and a simpler authoring environment) if they could simply tell StoreBroker to use a different language's screenshots when language-specific ones are not available.

With this change, the following end-user features are now enabled:
  * The StoreBroker config file has been updated to support a `MediaFallbackLanguage` config option
    that allows users to specify a lang-code (e,g, `en-us`, `es-es`, etc...)
  * Users can specify a `-MediaFallbackLanguage` parameter to `New-SubmissionPackage` and
    `New-InAppProductSubmissionPackage`, specifying a lang-code (e,g, `en-us`, `es-es`, etc...) which will
    override any config file option that has been specified.
  * Users can specify a `FallbackLanguage` attribute on the `<caption />` and `<icon />` elements in
     PDP files to optionally choose a different fallback language on a per-element basis.

In any of these cases, if the language-specific media file cannot be found, and a fallback language has
been specified, StoreBroker will then look in the fallback language's Images/media sub-folder for the
same-named media file.  Only then, if still not found, will packaging fail.

It should be noted that there can only ever be _one_ designated fallback language for any given file.
The logic does not cascade.  Therefore, if a user specifies a fallback language in both the config file
_and_ at the command prompt, the command prompt always wins.  If a value has been specified via
the config file and/or command prompt, _as well as_ in a PDP file, the fallback language from the PDP
file will be the one that is used _for that element_, and then StoreBroker will revert back to using the
one from the config file / command prompt until it encounters another one defined specifically in
the PDP file.

Addresses #28: Unable to submit common screenshot for all languages