Capture activityId to log with 500 / Internal Server Error failures
According to the API team, in the cases where they return back a
500 / Internal Server Error, they will return back JSON content
in the body that contains an `activityId` property, which will be
a GUID.  This property will apparantly help the Submission API team
diagnose these server errors, so, we now try to log it.

A `WebException` contains an `HttpWebResponse` object which requires
stream access in order to read any embedded content.  So, I've added
a new helper method, `Get-HttpWebResponseContent` which we'll call
when accessing the `WebException` in order to get to the raw content.

If we are able to get content, deserialize it, _and_ there's an
`activityId`, we'll just log that `activityId`.  Otherwise, we'll
log the entire raw content (if it exists).

I'm also reverting a small part of change #6c8bed398, which attempted
to throw a custom error message containing the `MS-CorrelationId` in
a scenario where we would never actually have it.  In this case, I'm
simply reverting that code to just continuing to throw the original
exception.

This also fixes any dangling PSScriptAnalyzer issues that _can_ be fixed,
and explicitly notes in the code the ones that can't be fixed
due to (referenced) PSScriptAnalyzer bugs.