Prevent checks for global vars from going into $global:Error
The check for the existence of a global variable will by default
cause an innocuous error to wind up in `$global:Error` if the
variable doesn't exist.

We had been using `-ErrorAction SilentlyContinue` so that the user
didn't see those errors, but they still ended up in `$global:Error`.
Changing to `-ErrorAction Ignore` prevents the user from seeing
the error, _and_ prevents the error from ending up in `$global:Error`.