Fix for trying to get the RawStream for errors.
36384212b introduced a change to try to get additional error
content from the response stream when we received an error from
the server.

Unfortunately, there was a typo in the parameter name we were
attempting to access the content stream from, resulting in
a new exception being thrown: "You cannot call a method on a
null-valued expression".

This fixes the typo, and also makes the attempt to get that value
best-effort by doing a try/catch around the call, simply so that
we never lose reporting the true error from the API.