Fix unformatted error message when OutPath is not specified
`Resolve-PackageParameters` has a default error message that requires
parameter formatting to be applied to it to indicate the name of the
parameter that is missing.

While it's used correctly for `OutName`, it's used incorrectly for
`OutPath`.  `$out` is already joined with newlines and only needs to
be formatted.  `$OutName` does this correctly.

Fixing `$OutPath` to leverage `$out` the same way as `$OutName`.
Specifically for `OutPath` in