Provide helpful error message if users try to manipulate packages without specifying any
If a user tries to use `-AddPackages` or `-ReplacePackages` with a StoreBroker payload
that doesn't contain any, they end up with a literal `null` entry in the `applicationPackages`
list due to how we manipulate the submission object.  This then results in an error with the API.

We should actually be catching this scenario during patching time, and providing users a helpful
error message letting them know that they should check their input parameters to `New-SubmissionPackage`
so that they know how to recover.