Re-enable support for non-release structured ImagesRootPath
The recent addition of fallback media support assumed that
users would always specify a `Release` attribute within the PDP
file (and therefore made the Release parameter to `Get-LocalizedMediaFile`
a required attribute).  It turns out that we have users that aren't
specifying a value for the `Release` attribute in the PDP file,
and so were therefore hitting a parameter validation error while
running packaging.

This fix removes the `Mandatory` attribute from the `Release` parameter
to `Get-LocalizedMediaFile` in order to re-enable this usage scenario.