Avoid throwing an exception for duplicate images.
When fallback image support was implemented, we switched to throwing an exception for duplicate images. This is actually a breaking change for users that previously had an incorrect image folder structure.  We published the change using incorrect NuGet version semantics which makes it difficult for users to rollback to the non-breaking change.  To fix, we should change to only report a warning in this case.  This will allow users with incorrect image folder structure to still successfully package as they were before. In the future, we will switch back to the exception during packaging, but we will publish the change as a new major version because it is a breaking change.