Add auto-retry logic for configurable list of error codes
There is a set of known error codes where users can usually achieve success if they simply retry enough times:

  * `429` - The Submission API limits a Tenant to 20 requests per minute.  Any requests exceeding
       that will receive this error code until that minute has expired.

  * `503` - An underlying service component was unavailable, and the Submission API suggests that
       you try again.

This change modifies the default execution behavior of StoreBroker to auto-retry up to **5** times with exponential back-off for these two error codes.  The implementation adds two new global variables to permit users to further configure how this works:

 **`$global:SBAutoRetryErrorCodes`** - an `int` list of status codes that we should retry for
**`$global:SBMaxAutoRetries`** - an `int` indicating the maximum number of times we should retry before failing.

We will only report back telemetry when the retry count is reached (and we finally error out), or when the result is finally successful.  We are adding two new telemetry properties that will be reported for both (eventual) successful requests,, as well as for eventual exceptions:
   * `NumRetries` - to track the number of times we had to attempt a retry
   * `RetryStatusCode` - to track the last status code that was hit that caused us to need to do the retry.

Resolves #92