- Fix broken logging when -Message array includes a format operation.
After the recent changes to Write-Log, logging was broken for specific cases where -Message was an array
and included a format operation. The issue was our misunderstanding of the functionality of the format operator.
The operator can operate with an array on the left-hand side. In this case, the elements of the array are joined
and each string in the array may be formatted. When using the format operation while trying to create an array,
we were inadvertently joining the contents of the array into a single string when we actually intended the format
operation to be restricted to one line of the array.

The fix is to wrap the format operation in parentheses, to make the expression explicit. In the future, we should be
more diligent about using parentheses for the same effect.