Update CHANGELOG.md
CHANGELOG.md hasn't been updated since StoreBroker was first released.

We should change that.

I've already gone through and added the appropriate [`git tags`](https://git-scm.com/docs/git-tag)
for every release, and now the CHANGELOG will provide a very quick summary of every change as well
for quick and easy perusal.

Going forward, we should keep both this CHANGELOG and the tags up-to-date.

Additionally updated CONTRIBUTING.md to explain the procedure going forward,
and to call out all of the contributors that we've had so far.