Fix exception seen when running ConvertFrom-ExistingSubmission with no trailers
While this scenario had been tested brefore, clearly something had been missed.
It appears that if you pass `$null` (the result of $Submission.trailers` when
the application doesn't have Advanced Listing support) to `ForEach-Object`,
that iterates once (for the `$null`) which will eventually result into a call of
`$null | Get-Member -Type NoteProperty` which throws an exception since you haven't
passed an object to `Get-Member`.

The fix is to simply switch to a standard `foreach` as opposed to the `ForEach-Object`
cmdlet.