All apps support trailers and gaming options now
The Store has opened up gaming options and trailers to
all apps, whether or not they explicitly have the
"advanced listings" option added.

The API is making a change to always return back
`true` for `hasAdvancedListingSupport`, but that
doesn't completely unblock StoreBroker.  StoreBroker also
followed the guidance in the documentation that indicated
if `trailers` or `gamingOptions` was `null` in the cloned
submission, it meant that they weren't supported.  That
is no longer accurate and the MSDN documentation is being
updated, but that also means that StoreBroker needs to make
some minor changes to accomodate that change in behavior
from the API.