Migrate `New-ApplicationSubmission` to use `isMinimalResponse`
The Store API recently introduced a new option when cloning a new
Application submission (but not for IAP or Flight submissions):
    `isMinimalResponse=true`

When provided, the returned submission object will be shallow (missing
application packages and listings).  To get the full submission object,
you need to grab the `id` from that returned submission object and then
do a `GET` on that submission.

The benefit of doing this is reducing the likelihood that the API will
time out while trying to put together the newly cloned submission object
to return to us.  What has been happening is that when trying to clone
larger submissions, the service will return back a `500` error after it
has successfully cloned, but before it can return back the newly cloned
object.