Increase the potential spread for the retry logic
We knew already that we wanted to use a random value for the
intial sleep time on the retry in order to ensure a nice time
spread for all clients (to reduce the likelihood of continued
contentions).  However, to achieve this, we had PowerShell
randomly select one of five potential initial sleep times
(1, 1.25, 1.5, 1.75, 2).

To achieve a better spread, we'll instead let it truly select a random value between a larger initial spread of values between
0.40 and 2.0 with a larger range of granularity.

This change is attempting to solve a real problem we're
experiencing lately with an increased number of 429 errors
(too many requests in a given minute).