Add Load Balancing / Round Robin Support to StoreBroker Proxy
The Store API has two different rate limiters in place
with corresponding error codes:

429 - API rate limiter - a ClientId can't make more than
      60 queries per minute
403 - API quota limiter - a ClientId can't make more than
      60,000 queries per day

For organizations that have many StoreBroker clients,
it is quite possible that 429 errors may happen fairly
frequently (along with the occasional 403's on really bad
days).

To avoid this, the Proxy is gaining the concept of
load balancing (from the API/Client Id perspective) by
enabling administrators to configure a TenantId to have
as many ClientId's as they'd like -- the Proxy will use
a "round robin" approach and simply cycle through all
configured ClientId's for a given TenantId/EndpointType
combination upon each request.

As more StoreBroker clients are added within an organization,
the Proxy administrator can simply add in additional ClientId's
as needed.  This change will be completely transparent to all
clients, and help ensure reliable API execution that doesn't
hit any rate or quota limiters.

Resolves Issue #114 - Add ClientID round-robin support to StoreBroker Proxy

Also, some additional headers need to be proxied to the API.
Clients are able to provide the "MS-CorrelationId" and "MS-Client-RequestId"
headers in their requests, and when specified, these should be proxied
directly to the API.

We also want to return back any header that includes the word
"retry" in it, as the API may start to use those to indicate how
long we should wait before retrying the same request.

Finally, we want to provide a default value for the return message's
ContentType in the event that the API is incorrectly returning back
a message without one specified (in this scenario, we'll default to
`text/plain`) (an issue that has occurred in the past).