Migrate New-ApplicationFlightSubmission to use isMinimalResponse
This is a follow-up to cdee54f6e3a23be29bf7f75df88af9fa4d4a648f (Pull Request #124).
The API team originally added `isMinimalResponse=true` as an optional parameter to the
clone/POST command for new submissions, which had it return a sparse/minimal response
in order to reduce the likelihood of getting a `500` error response while it tried to
gather the complete data for the full submission response.

They have now added support for the same scenario for Flights as well, and so the same
type of change is being made to `New-ApplicationFlightSubmission` so that it creates
the cloned submission with `IsMinimalResponse=true` and then immediately tries to
`GET` that submission to get the full content.