StoreBrokerProxy should proxy the Location header from the real response
For API v2, the API will start returning a Location header that clients
should honor for subsequent requests.  This follows the
[HTTP/1.1 standard](https://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html#sec9.5)
which states:

>if a `POST` operations results in the creation of a resource, then the reponse
> should include a `Location` header with the address of the new resource.

The API will be using the `Location` for other scenarios as well, but in order for
StoreBroker to leverage it, we need to ensure that the Proxy is returning the original
value.