Protect against formatting invalid certification reports
Per the [documentation](https://docs.microsoft.com/en-us/windows/uwp/monetize/manage-app-submissions#certification-report-object),
a certification report is supposed to have both a `date` and `reportUrl` properties.

We've seen instances where the API returns back certification reports that don't have a `date` property, and where the `reportUrl`
contains the message `Obsolete: please visit https://developer.microsoft.com for report details.`.  When this happens, we end up
throwing an exception as we attempt to format a `$null` date value.

This update protects against that scenario by ensuring that we only attempt to format the date if one is provided.