Don't reference System Environment folders during API interaction since servers made not have them
During `Update-Submission`, if we detected that a user's JSON was missing the `ProductId`, we
printed a message to them informing them how to create a new config file.  Part of that message
involved putting the new config file on their desktop using `[System.Environment]::GetFolderPath('Desktop')`.

Unfortunately, when running in a server environment, that will return back `$null`, and then when used
with `Join-Path`, that will result in an error for trying to join a `$null path`.  The fix here is to
just show a command where the new config file is being generated in the current folder.
We were trying to print a message to users informing them how to create a new config since theirs