Add support for a minimal input object.
For the Office scenario, they are often in a case where they
may only have a few updated strings across all of the possible
metadata that a listing may contain.

Unlike a PDP.xml-based update, Office will have an incomplete,
or minimal, input object.  If provided like that to StoreBroker
today, that would result in the deletion of any language not
contained in the minimal object, and all non-specified text
properties being set to null.

To avoid that, we're now introducing the concept of
`IsMinimalObject`.  If specified to `Update-Submission` (or to
a subset of the `Update-*` methods) then only properties that
were included in the input object will be set on the cloned
submission.

Updates to Images/Videos and their captions are explicitly
excluded from this behavior (and in fact, will be skipped
entirely in this mode) given that we can not accurately
and reliably determine what right thing to do would be given
that the same file content can be referenced by
differently-named files.

This also updates all `Update-*` methods to use an `Add-Member`
derived solution for setting properties on the returned object,
since the API does not guarantee that any given property will
exist on a returned object if its current value is `null`. I
say an "`Add-Member` derived solution", as I've created a
wrapper called `Set-ObjectProperty` which can _also_ test
if a property exists on a source object before attempting
to set the property on the `InputObject` (which is the scenario
we're trying to solve for in the minimal object case).