Remove `WaitSeconds`
The `WaitSeconds` `GET` parameter for a new submission has been deprecated
by the Store API team in favor of using the `202`/`200` approach to indicate
to clients that they need to re-query until the submission is ready.

Remove `WaitSeconds` and update telemetry usage to log the `WaitUntilReady`
and `WaitForCompletion` switch usage.