Fixing switch parameter name for Patch-ApplicationFlightSubmission
All of the "flight" methods expected a switch called `-UpdatePublishMode`,
but in `Patch-ApplicationFlightSubmission`, it was actually called
`UpdatePublishModeAndVisibility`.  Due to PowerShell's logic of doing
best-effort parameter matching when there's no ambiguity, the existing logic
worked fine since it was always called as `-UpdatePublishMode` (which is a strict
substring of `-UpdatePublishModeAndVisibility).

To remove any potential confusion for code inspectors however, fixing this
accidental typo.

Resolves #143