Normalize invocation logging
Created helper method `Write-InvocationLog` which handles how we log the
main entrypoints into the module.  This will now log the actual values that
the user provided, as opposed to whatever the commandline looked like at
invocation time (with variable references, etc...).

All API accessing methods allow users to provide an AccessToken to be used
for the duration of that API call.  The problem with that is that the
command is also logged by default, which means that the AccessToken value
might be logged in plain text to the log file.

To fix this:
  * Callers of `Write-InvocationLog` can redact the value of specified parameters,
    or exclude the parameter altogether.
  * `AccessToken` has been configured to _always_ be redacted, and
    `NoStatus` has been configured to _always_ be excluded (to avoid noise)