Grab appxbundle/appx manifests by direct path instead of recursive search.
Previously, we would recursively search under the bundle/appx root for the desired manifest. This caused problems for a customer whose packaging scripts had placed an unrelated manifest in a subdirectory of the appx.

In a valid bundle/appx, the manifests are expected to be at a specific relative path. This change updates PackageTool to acquire the manifests from the known path.