Allow clients to optionally configure a ClientName that _they_ can send with telemetry
Added a new optional global configuration property `$global:SBStoreBrokerClientName` that
users can optionally specify to send along a name for their tooling that is consuming
StoreBroker, to make it easier for us to better understand how StoreBroker is being used.

This was initially added to make it easier to track usage coming from the public
Azure DevOps Build Extensions built on top of StoreBroker v2.

If specified, the client name will be reported in the telemetry for the following function calls:
 * `Update-Submission`
 * `New-SubmissionPackage`
 * `Update-SubmissionRollout`