Fix occasional error "ConvertFrom-Json : Invalid JSON primitive: Index"
This was happening while trying to access a header that may not have
been there in a response.  Specifically, the RetryAfter header value
doesn't exist in some situations (like authentication failure), and
trying to access caused an error cascade.

The proper fix is to verify that the header value exists before
querying for it.