Get New-StoreBrokerConfig working
New-StoreBrokerConfig can now generate a valid config file.
It works with or without a ProductId being specified.

If a ProductId is specified, it will fill that value into the
appSubmission.ProductId field, but it does not yet look up the recent
submission and populate the remainder of the appSubmission config values.