Update Resolve-ProductId to be able to resolve ProductId from the config
Previously, this was only ever tested against config files that had been
migrated from v1 over to v2 (meaning that they had appId's present).

As of 2.0.7, StoreBroker can create v2 config files, but that means that
it won't have the AppId, which means that Resolve-ProductId needs to also
support getting the ProductId from the config file as well.