Fix logic for adding new localized language listings
The logic in `Update-Listing` incorrectly treated the `listings`
property on `SubmissionData` like an array (checking for the `Count` of
elements it contains).  The languages are actually treated as _properties_
on the object, and thus asking for the `$SubmissionData.listings.Count` always
returned back `$null` (and iterating on it did nothing).

Instead, we need to be looking specifically at its `NoteProperty` members,
and iterating on, or checking the count of, _those_.