Fix Open-Store support for AppId
`Open-Store` uses the `Get-ProductStoreLink` to determine the
web and Store App URI's which requires a `ProductId`.  `Open-Store`
supports being called directly with a `ProductId`, or instead
with an `Appid` (it will simply use that to look up the `ProductId`
via `Get-Product`).  There was, however, an incorrect check was
determining if there was a valid `ProductId` by comparing it to `$null`
instead of checking if it was also empty.

Because of that incorrect check, we ended up calling `Get-ProductStoreLink`
with an empty `ProductId` which caused the validation logic in
`Get-ProductStoreLink` to return the error message that a user likely
passed-in an AppId and should therefore call it with the `-AppId` parameter.

That message was confusing, as it's the same validation error that `Open-Store`
would report back to users if they passed in an `AppId` as a `ProductId`, and yet
in this case the user _had_ actually passed in `-AppId`.

The fix here was to simply replace the `$null` check with a `[String]::IsNullOrWhiteSpace`
check.