Change server-requested retry logic from recursion to looping
Previously, if a request returned a `202` response (result not yet ready),
`Invoke-SBRestMethod` would sleep the requested amount of time, and then call
recursively call itself to try again.

This implementation was perfectly reasonable, except that some API commands
have been increasing in how long they take to complete, and the result has been
large recursive call stacks that are too much for the system to handle:
`The script failed due to call depth overflow.`

To fix this, I have migrated the handling of `202` results from using recursion
to taking advantage of the loop that already exists to handle the other retry
logic that `Invoke-SBRestMethod` supports for other response codes: instead of
using recursion, it simply updates `$UrlFragment` with the new location from the
header, and then uses `continue` to start the loop all over again.  Since the
same stopwatch will be used for the next telemetry event in the loop, we also
reset the stopwatch before continuing so that each retry has an accurate duration.